+++ .vscodeignore
@@ -5,9 +5,15 @@
 
 # 源代码文件（已编译到out目录）
 src/**
-media/**
 **/*.ts
 **/*.map
+
+# Webview源代码（已构建到dist目录）
+webview/src/**
+webview/public/**
+webview/node_modules/**
+webview/package-lock.json
+webview/vite.config.js
 
 # 配置文件
 .gitignore
@@ -16,21 +22,34 @@ vsc-extension-quickstart.md
 **/tsconfig.json
 **/.eslintrc.json
 vite.config.js
-
-# 示例文件（用户不需要）
-exampleFiles/**
+config/**
 
 # 开发依赖
 node_modules/**
 package-lock.json
 
-# 文档和说明
-README.md
-CHANGELOG.md
+# 文档和说明（保留在扩展中）
+# README.md
+# CHANGELOG.md
 
 # 测试文件
 **/test/**
+out/test/**
 
-# 临时文件
+# 临时文件和缓存
 **/*.log
 **/*.tmp
+.cache/**
+.vite/**
+
+# 构建工具和脚本
+tools/**
+
+# 开发文档
+BUILD_VERIFICATION.md
+REFACTOR_PLAN.md
+REFACTOR_SUMMARY.md
+
+# Git相关
+.git/**
+.github/**
