+++ examples/0_3840X2160_14_RGGB_seq_idx_channel_23309_0_0.raw
