+++ examples/0_640X480_10_RGGB_seq_idx_channel_72556_0_0.raw
