+++ examples/Baboon_512x512_8bit.raw
