+++ examples/Barbara_512x512_8bit.raw
