+++ examples/Cameraman_512x512_8bit.raw
