+++ examples/Goldhill_512x512_8bit.raw
