+++ examples/Lena_512x512_8bit.raw
