+++ examples/Peppers_512x512_8bit.raw
