+++ examples/w2688_h1520_10bit_RGGB.raw
