--- media/index.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Raw Image Viewer (Vue)</title>
-</head>
-
-<body>
-  <div id="app"></div>
-  <script type="module" src="./main.js"></script>
-</body>
-
-</html>
