+++ webview/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "raw-image-viewer-webview",
+  "private": true,
+  "version": "0.0.4",
+  "type": "module",
+  "scripts": {
+    "dev": "vite --config ../vite.config.js",
+    "build": "vite build --config ../vite.config.js",
+    "preview": "vite preview --config ../vite.config.js"
+  },
+  "dependencies": {
+    "vue": "^3.4.27",
+    "pinia": "^3.0.3"
+  },
+  "devDependencies": {
+    "@vitejs/plugin-vue": "^5.0.5",
+    "vite": "^5.4.19"
+  }
+}
