+++ webview/src/components/controls/ControlsPanel.vue
@@ -104,7 +104,7 @@
 
 <script setup>
 import { ref, computed, watch } from 'vue';
-import { useImageStore } from '../stores/image';
+import { useImageStore } from '../../stores/image';
 import { storeToRefs } from 'pinia';
 
 const store = useImageStore();
